+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -95,6 +95,11 @@ New checks
   Finds virtual classes whose destructor is neither public and virtual nor
   protected and non-virtual.
 
+- New :doc:`hooks-guard-in-for
+  <clang-tidy/checks/hooks-guard-in-for>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`misc-hook-entry-points
   <clang-tidy/checks/misc-hook-entry-points>` check.